+++ src/processing.py
@@ -1,19 +1,21 @@
-def filter_by_state(operations: list[dict], state: str = "EXECUTED") -> list[dict]:
+from typing import List, Dict
+
+
+def filter_by_state(operations: List[Dict], state: str = "EXECUTED") -> List[Dict]:
     """
     Фильтрует список операций по значению ключа 'state'.
-    :параметр operations: Список словарей с данными о банковских операциях.
-    :параметр state: Значение для фильтрации по ключу 'state'. По умолчанию 'EXECUTED'.
+    :param operations: Список словарей с данными о банковских операциях.
+    :param state: Значение для фильтрации по ключу 'state' ("EXECUTED" или "CANCELED"). По умолчанию "EXECUTED".
     :return: Отфильтрованный список словарей.
     """
     return [operation for operation in operations if operation.get("state") == state]
 
 
-def sort_by_date(operations: list[dict], reverse: bool = True) -> list[dict]:
+def sort_by_date(operations: List[Dict], reverse: bool = True) -> List[Dict]:
     """
     Сортирует список операций по дате.
-    :параметр operations: Список словарей с данными о банковских операциях.
-    :параметр reverse: Если True, сортировка по убыванию (по умолчанию). Если False,
-    :по возрастанию.
+    :param operations: Список словарей с данными о банковских операциях.
+    :param reverse: Если True, сортировка по убыванию (по умолчанию). Если False — по возрастанию.
     :return: Отсортированный список словарей.
     """
     return sorted(operations, key=lambda x: x["date"], reverse=reverse)
@@ -34,12 +36,25 @@ data_date = [
     {"id": 615064591, "state": "CANCELED", "date": "2018-10-14T08:21:33.419441"},
 ]
 
-# Фильтрация по состоянию
-filtered_data = filter_by_state(data_state, state="EXECUTED")
-print("Отфильтрованные данные:")
-print(filtered_data)
 
-# Сортировка по дате
-sorted_data = sort_by_date(data_date, reverse=True)
-print("Отсортированные данные:")
-print(sorted_data)
+# Примеры использования
+if __name__ == "__main__":
+    # Фильтрация по "EXECUTED" (используется data_state)
+    executed_operations = filter_by_state(data_state, state="EXECUTED")
+    print("Отфильтрованные операции (EXECUTED):")
+    print(executed_operations)
+
+    # Фильтрация по "CANCELED" (используется data_state)
+    canceled_operations = filter_by_state(data_state, state="CANCELED")
+    print("Отфильтрованные операции (CANCELED):")
+    print(canceled_operations)
+
+    # Сортировка по убыванию (используется data_date)
+    sorted_desc = sort_by_date(data_date, reverse=True)
+    print("\nОтсортированные операции (по убыванию):")
+    print(sorted_desc)
+
+    # Сортировка по возрастанию (используется data_date)
+    sorted_asc = sort_by_date(data_date, reverse=False)
+    print("\nОтсортированные операции (по возрастанию):")
+    print(sorted_asc)
